Extract ConversionHelpers to separate files.
